Refactor: break down utils.
