+++ src/explorer/model/MavenProject.ts
@@ -8,6 +8,7 @@ import { Settings } from "../../Settings";
 import { getPathToExtensionRoot } from "../../utils/contextUtils";
 import { Utils } from "../../utils/Utils";
 import { mavenExplorerProvider } from "../mavenExplorerProvider";
+import { EffectivePom } from "./EffectivePom";
 import { ITreeItem } from "./ITreeItem";
 import { MavenPlugin } from "./MavenPlugin";
 import { PluginsMenu } from "./PluginsMenu";
@@ -16,13 +17,13 @@ const CONTEXT_VALUE: string = "MavenProject";
 
 export class MavenProject implements ITreeItem {
     public parent?: MavenProject;
-    private _rawEffectivePom: string;
-    private _effectivePom: any;
+    private _effectivePom: EffectivePom;
     private _pom: any;
     private _pomPath: string;
 
     constructor(pomPath: string) {
         this._pomPath = pomPath;
+        this._effectivePom = new EffectivePom(pomPath);
     }
 
     public get name(): string {
@@ -37,25 +38,21 @@ export class MavenProject implements ITreeItem {
         return _.get(this._pom, "project.modules[0].module") || [];
     }
 
-    public get rawEffectivePom(): string {
-        return this._rawEffectivePom;
-    }
-
-    public get effectivePom(): string {
+    public get effectivePom(): EffectivePom {
         return this._effectivePom;
     }
 
     public get plugins(): MavenPlugin[] {
         let plugins: any[];
-        if (_.get(this._effectivePom, "projects.project")) {
+        if (_.has(this._effectivePom.data, "projects.project")) {
             // multi-module project
-            const project: any = (<any[]>this._effectivePom.projects.project).find((elem: any) => this.name === _.get(elem, "artifactId[0]"));
+            const project: any = (<any[]>this._effectivePom.data.projects.project).find((elem: any) => this.name === _.get(elem, "artifactId[0]"));
             if (project) {
                 plugins = _.get(project, "build[0].plugins[0].plugin");
             }
         } else {
             // single-project
-            plugins = _.get(this._effectivePom, "project.build[0].plugins[0].plugin");
+            plugins = _.get(this._effectivePom.data, "project.build[0].plugins[0].plugin");
         }
         return this._convertXmlPlugin(plugins);
     }
@@ -92,30 +89,27 @@ export class MavenProject implements ITreeItem {
     public getChildren(): vscode.ProviderResult<ITreeItem[]> {
         const ret: ITreeItem[] = [];
         ret.push(new PluginsMenu(this));
-        if (this.moduleNames.length > 0 && Settings.viewType() === "hierarchical" ) {
+        if (this.moduleNames.length > 0 && Settings.viewType() === "hierarchical") {
             ret.push(...this.modules.map(m => mavenExplorerProvider.getMavenProject(m)));
         }
         return ret;
     }
 
-    public async calculateEffectivePom(force?: boolean): Promise<void> {
-        if (!force && this._rawEffectivePom) {
-            return;
+    public async calculateEffectivePom(force?: boolean): Promise<string> {
+        if (!force && this._effectivePom.raw) {
+            return this._effectivePom.raw;
         }
 
-        this._rawEffectivePom = await Utils.getEffectivePom(this);
-        await this._parseEffectivePom();
+        await this._effectivePom.update();
         mavenExplorerProvider.refresh(this);
-    }
-
-    public async refreshPom(): Promise<void> {
-        await this.parsePom();
-        mavenExplorerProvider.refresh(this);
+        return this._effectivePom.raw;
     }
 
     public async refresh(): Promise<void> {
-        await this.refreshPom();
-        this._rawEffectivePom = undefined;
+        await this._refreshPom();
+        this._effectivePom.update().then(() => {
+            mavenExplorerProvider.refresh(this);
+        }); // no await to unblock the thread.
     }
 
     public async parsePom(): Promise<void> {
@@ -126,12 +120,9 @@ export class MavenProject implements ITreeItem {
         }
     }
 
-    private async _parseEffectivePom(): Promise<void> {
-        try {
-            this._effectivePom = await Utils.parseXmlContent(this._rawEffectivePom);
-        } catch (error) {
-            this._effectivePom = undefined;
-        }
+    private async _refreshPom(): Promise<void> {
+        await this.parsePom();
+        mavenExplorerProvider.refresh(this);
     }
 
     private _convertXmlPlugin(plugins: any[]): MavenPlugin[] {
