+++ src/utils/contextUtils.ts
@@ -4,14 +4,17 @@
 import * as fse from "fs-extra";
 import * as os from "os";
 import * as path from "path";
-import { ExtensionContext,  extensions } from "vscode";
+import { ExtensionContext, extensions } from "vscode";
 
+let EXTENSION_CONTEXT: ExtensionContext;
 let EXTENSION_PUBLISHER: string;
 let EXTENSION_NAME: string;
 let EXTENSION_VERSION: string;
 let EXTENSION_AI_KEY: string;
 
 export async function loadPackageInfo(context: ExtensionContext): Promise<void> {
+    EXTENSION_CONTEXT = context;
+
     const { publisher, name, version, aiKey } = await fse.readJSON(context.asAbsolutePath("./package.json"));
     EXTENSION_AI_KEY = aiKey;
     EXTENSION_PUBLISHER = publisher;
@@ -45,4 +48,9 @@ export function getPathToTempFolder(...args: string[]): string {
 
 export function getPathToExtensionRoot(...args: string[]): string {
     return path.join(extensions.getExtension(getExtensionId()).extensionPath, ...args);
+}
+
+export function getPathToWorkspaceStorage(...args: string[]): string {
+    fse.ensureDirSync(EXTENSION_CONTEXT.storagePath);
+    return path.join(EXTENSION_CONTEXT.storagePath, ...args);
 }
